Move ID duplication checks to different method
